Rename class Api\Data\OmiseChargeInterface to Api\PaymentMethodManagementInterface and all of classes that related with this change
