+++ Api/PaymentMethodManagementInterface.php
@@ -1,0 +1,6 @@
+<?php
+namespace Omise\Payment\Api;
+
+interface PaymentMethodManagementInterface
+{
+}
