+++ etc/webapi.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Webapi:etc/webapi.xsd">
     <route url="/V1/order/:orderId/payment-authorize-uri" method="GET">
-        <service class="Omise\Payment\Api\Data\OmiseChargeInterface" method="getAuthorizeUri"/>
+        <service class="Omise\Payment\Api\PaymentMethodManagementInterface" method="get3DSecureAuthorizeUri"/>
         <resources>
             <resource ref="anonymous" />
         </resources>
